requirements.txt
# UI
PyQt6>=6.4.0
PyQt6-Qt6>=6.4.0
PyQt6-sip>=13.4.0
qt_material==2.14

# Video Processing
opencv-python>=4.8.0
python-vlc>=3.0.20000

# Voice Recognition
vosk>=0.3.45
pyaudio>=0.2.13

# Event System
pypubsub>=4.0.3
watchdog>=3.0.0

# QR Code
qrcode>=7.4.2
pillow>=10.0.0

# Development Tools
pytest>=7.4.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0

# Packaging
pyinstaller>=5.13.0 